Workaround for GNU Autoconf unescaped left brace
See [perl #130497]

GNU Autoconf depends on Perl, and will not work on Blead (and the
forthcoming Perl 5.26), due to a single unescaped '{', that has
previously been deprecated and is now fatal.  A patch for it has been in
the Autoconf repository since early 2013, but there has not been a
release since before then.

Because this is depended on by so much code, and because it is simpler
than trying to revert to making the fatality merely deprecated, this
patch simply changes perl to not die when compiled with the exact
pattern that trips up Autoconf.  Thus Autoconf can continue to work, but
any other patterns that use the now illegal construct will continue to
die.  If other code uses the exact pattern, they too will not die, but
the deprecation message continues to get raised.  The use of the left
brace in this particular pattern is not one where we envision using the
construct to mean something else, so a deprecation is suitable for the
foreseeable future.