add SHORT PVs
This commit adds an option for storing small strings (<16 bytes on a
64-bit system) directly in the XPV body of an SV.

It's still work-in-progress. It passes all tests, but could probably
benefit from optimisation work.

It's not enabled by default: perl must be built with the new cpp define
-DPERL_COPY_ON_WRITE3.

As the name implies, this commit is the first part of what's intended to
be a revamp of the COW implementation (COW1 was Nicholas's, COW2 was
sprout's; this is COW3).

The main things this commit does is:

1) Adds an extra pointer-sized field to the XPV structure and all those SV
   types which inherit from it; so there's now (xpv_cur,xpv_len,extra)
   instead of just (xpv_cur,xpv_len) in every string-capable body type.

2) Adds a new SV flag, SVf_SHORTPV, which indicates that the string is
   stored directly in the body.

Under SVf_SHORTPV, sv_any points to the body as normal, but sv_u.svu_pv
points into the middle of the same body, where the (xpv_len,extra) fields
are used as a 16-byte buffer. xpv_cur still holds the current string
length.

The SvLEN() macro has been changed to return a constant value if SVf_SHORTPV.
The SvCUR() and SvPVX() macros are unchanged.

Since derived string types like XPVIV also have the extra field, they can
hold a short string as well as an integer value.

When upgrading a SVf_SHORTPV SV to a new body type, the old body isn't
discarded but remains as the SV's string buffer, so sv_any and sv_u.svu_pv
now point to two different bodies. This means that code calling sv_upgrade
doesn't see the buffer address change. The old body will be freed when the
SV is freed.

Note that nothing under ext/, dist/ or cpan/ needed fixing apart from
ext/PerlIO-encoding/, which did some direct PVX buffer stealing.
(And Peek.t, obviously.)

The main breakages in core itself were places that directly moved a PVX
buffer from one SV to another: typically in the implementation of matching
and substitution - places such as pp_subst.