sv_chop() may reallocate buffer or shift bytes
Normally sv_chop() just uses the OOK mechanism to logically remove bytes
from the beginning of a string without physically moving anything.
However, it's possible for the sv_chop() to reallocate the buffer, so on
return the ptr arg and the original SvPVX(sv) may no longer point to
anything valid.

Later on in this branch it is hoped to introduce SVf_SHORTPV strings,
which sv_chop() will operate upon by moving bytes rather than activating
the OOP mechanism. This will make invalid ptr/SvPVX(sv) more common.

This commit does two things in preparation.

First, it tweaks the perlapi pod for sv_chop(): a comment about not
relying on the value SvPVX has been clarified;

second, a couple of callers of sv_chop in the perl core have been fixed
to not assume an invariant SvPVX buffer.