Add more extensive information about @INC changes into perldelta (RT#131304)
This mainly takes existing text written by Dave Mitchell, Kent Fredric,
Sawyer X and James E Keenan.

[Ævar: I applied this from RT #131304. It had conflicts which I
resolved to the best of my ability.

The submitted patch removed the "'.' and @INC" section which I rewrote
in 90ed3b64f9 ("perldelta: clarify what '.' and @INC is about",
2017-05-16). I resolved the conflict to keep it.

Maybe this was the wrong move. We're now mentioning the removal of "."
in 3 prominent places, at the top in "Notice", "Core Enhancements",
and "Security".

Maybe the best move here is to remove the section I wrote, although as
noted in my commit I think one redeeming factor it if is the plain
language describing the change (i.e. not assuming knowledge of
variables like @INC right off the bat), but obviously I'm biased
there].