perldelta: Cleanup documentation changes section
I looked at all the files in /pod for their changes since 5.24.  I think
this action should be added to the RMG.  I removed the changes that are
described elswhere in perldelta, adding boiler plate to that effect.
I removed the changes I thought would not be useful for someone reading
this, but added significant numbers of substantive changes that had been
omitted.