perldelta: update @INC blurb in Notice
The original text got mangled when applied; I've readded the missing
sentence, and reworded things slightly to be a more informative initial
mention of what the issue is (like mentioning that the change is for
security reasons).

This should hopefully cover anything that was missed in the text removed
by 2c4393cceb72ac358b.