perldelta: remove some "Internal Changes" entries
In general we don't list internals changes unless it affects the API or is
very visible (e.g. performance enhancement). So I've deleted the following
entries:

    Several new internal C macros have been added that take a string
    literal as arguments, alongside existing routines that take the
    equivalent value as two ...

AFAKT, this is referring to:

    strEQs() and strNEs()
    _memEQs() and _memNEs

which aren't listed as part of the API, so we don't (yet?) want to
advertise them.

    The code in F<gv.c> that determines whether a variable has a special
    meaning to Perl has been simplified.

which is great, but isn't visible AFAIKT.

    Use C<my_strlcat()> in C<locale.c>.  While C<strcat()> is safe in this
    context, some compilers were optimizing this to C<strcpy()> causing a
    porting test to fail that looks for unsafe code.  Rather than fighting
    this, we just use C<my_strlcat()> instead.

Perhaps this should be reported as bug fix instead?