perldelta: Collapse "Removed Deprecations"
I just added this section in 1f81bc188d3574dfda690659ebeed2bdc92a42d6,
but on further contemplation, I think it makes more sense to just have
the "Deprecations" section deal with anything regarding deprecations.
It is short, and undeprecating something is rare, and I have revised the
entry so it makes sense being in the plain "Deprecations" section.