perldelta: Standardize references to Perl
There is a mixture of saying Perl "5.foo" and "v5.foo".  This
standardizes to the two most common variants.  When the word "Perl"
precedes the version number, the "v" is omitted.  Perl is capitalized.
When the number stands alone, the "v" is added.

However, the NAME section must say 'perl v5.26.0' as that form is
expected from Porting/pod_lib.pl