perldelta: Fix link displays
This fixes the displays of some links.  Most of this fix is
that one shouldn't say

C<L<foo/bar>>

This is because it translates to 'bar in foo' entirely in the
fixed-width font.  And 'in foo' shouldn't be in fixed width.  (I will
bring this up to see if Pod::Simple should change, or it should be
flagged by Pod::Checker.)

Instead one should currently write

L<C<bar>|foo/bar>