utf8.h: Add parens around macro param in expansion
a6951642ede4abe605dcf0e94b74948e0a60a56b added an assertion to find bugs
in calling macros, and so far, instead, it found a bug in a macro.  A
parameter needs to be enclosed in parens in case it is an expression, so
that precedence works.

(cherry picked from commit 4e1ed312da261450ba45a56e7b6756a873678f52)