add t/porting/bench.t
This script tests the functionality of Porting/bench.pl - but currently
only on Linux, and only when valgrind and the .git/ dir are present.

(an earlier script of the same name has recently been renamed to
t/porting/bench_selftest.t.)

Its minimal at the moment; it will be added to as bugs in bench.pl are
addressed.