t/re/reg_mesg.t: Test that death actually happens.
Until now, patterns that are supposed to die weren't actually being
checked for that.  Instead the message that was raised was checked
against the expected value, and if they matched, it assumed that death
accompanied it.  But this isn't necessarily true.  Also, some entries
were given an empty expectation, which was wrong.  So this makes sure of
that.  Those empty patterns matched anything at all, so things went
undetected.

To be sure of death, this commit appends a statement with a lingering
effect to the string eval of what's supposed to die.  If the effect
happens, it didn't die.