Allow bitwise & ^ | to accept trailing UTF-8
Commit 08b6664b858b8fd4b5c0c27542763337b6d78e46 breaks things like

$foo = "" & "\x{100}";

We have deprecated using above-FF code points in bitwise operations, and
made them illegal in 5.27.  However, the case where the illegal code
points don't play a part in the operation never raised deprecation
warnings.  The example above is one such, because the \x{100} comes
after the operation stops since the other operand has length 0.

We can't make something illegal without warning people about it for 2
releases.

Rather than revert that commit, and reinstate a bunch of slow code that
is far more general than now needed, this commit adds some extra code to
deal with these situations, but the basic operations still take place in
tight loops, which 08b6664b858b8fd4b5c0c27542763337b6d78e46 caused to
happen.

In the case of "&", the illegal code points get truncated away.  In the
case of ^ and |, they get catenated as-is.  This preserves earlier
behavior.

It has not been decided if these should at least warn, or the usage
should be deprecated.  A commit can easily be done to change to whatever
the final decision is, but this commit doesn't raise any warnings, hence
preserves existing behavior.

The breaking commit looks like it might create some havoc with CPAN, and
fixing it now will save the CPAN testers effort, as they won't have to
deal with a bunch of broken distributions.