APItest/t/utf8_warn_base.pl: Skip most tests
This test file tests every end-of-Unicode-plane noncharacter, and a
middling surrogate, and a nonchar in the interior of the consecutive
range of them.  But, we don't really have to do more than basic testing
for these  middling cases.  We should test that they are detected as
being in their respective categories, but testing that all combinations
of warning and disallowed flags and return flags shouldn't be necessary.
It's sufficient to test for those for the real edge cases.

This cuts the number of tests in this file to somewhat less than 1/3 of
the original.