Porting/bench.pl ban duplicate labels in multiread
With multiple --read options, die if the same label appears twice.

A duplicate between a --read file and on the command line was already
checked for, but not from two reads.

Also, improve the error messages.

Finally, regenerate the callsub2.json  test file using a './perl2' label
this time, so that both callsub.json and callsub2.json can still be read
without error.