S_gen_constant_list(): initialise si_stack_hwm
This is purely to shut up valgrind.

S_gen_constant_list() directly calls Perl_pp_anonlist() while not
in a runops loop. So PL_curstackinfo->si_stack_hwm doesn't get
initialised. Which is harmless, as only a runops loop is concerned with
its value. But valgrind gets upset:

    Conditional jump or move depends on uninitialised value(s)
       at 0x651D90: Perl_pp_anonlist (pp.c:5257)

Can be reproduced with '@a = (1,0..0);' (or any list that can be constant
folded).